+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Modified existing request/response to accomodate dynamic rate hub request/response
 - Allowed rounding for length, width, and height
 - Modified the routing to connect with the dynamic rates hub
+- Added ISO Mapping for USA
+- Added zipcode trimming
 
 ### Fixed